Add implementation with Mercado Pago: PaymentService and PixService

+++ DigitalCommerceAPI/src/main/java/com/arthur/digitalcommerce/payload/ProductDTO.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -18,9 +20,9 @@ public class ProductDTO {
     private String description;
     private String image;
     private Integer quantity;
-    private double price;
-    private double discount;
-    private double specialPrice;
+    private BigDecimal price;
+    private BigDecimal discount;
+    private BigDecimal specialPrice;
     private Long categoryId;
